% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ani.barplot.R
\name{ani.barplot}
\alias{ani.barplot}
\title{Animated Barplot}
\usage{
ani.barplot(
  x,
  col = "lightblue",
  ylab = "",
  xlab = "",
  main = "",
  ylim = c(0, max(x) + max(x)/6),
  box = TRUE,
  names = "",
  cex.names = par("cex.axis"),
  border = "black",
  space = 0.4,
  text = TRUE,
  cex = 1,
  coltext = 1,
  pre = "n =",
  post = "\%",
  format = "png",
  digit = 2,
  frames = 100,
  folder = "",
  ani.text = TRUE,
  width = 1920,
  height = 1080,
  point = 8,
  res = 300,
  bg = TRUE,
  bg.from = "grey",
  bg.to = "white",
  bg.increase = TRUE,
  axes = TRUE,
  cex.axis = par("cex.axis")
)
}
\arguments{
\item{x}{a vector or one dimensional frequwncy table of values describing the bars which make up the plot.}

\item{col}{a vector of colors for the bars. By default col is set to "lightblue".}

\item{ylab}{a label for the y axis}

\item{xlab}{a label for the x axis}

\item{main}{title of plot}

\item{ylim}{limits for the y axis.}

\item{box}{a logical value. If FALSE no box around plot will be drawn}

\item{names}{a vector of names to be plotted below each bar. If this argument is left empty, the names are taken from the names freuqency table of x}

\item{cex.names}{expansation of names below bars}

\item{border}{color of bar border}

\item{space}{space between bars}

\item{text}{a logical value. If TRUE text will be drawn on bars}

\item{cex}{expansation of text on bars}

\item{coltext}{text color}

\item{pre}{pre text of absolute frequencies on bars}

\item{post}{post text of absolute frequencies on bars}

\item{format}{the image output format e.g.: "png", "pdf", "jpeg", etc.}

\item{digit}{the numer of decimal places to display relative frequencies with e.g.: digits=3 -> 22,2}

\item{frames}{animation length in frames}

\item{folder}{folder to save animation as png frames to. If "" no images will be saved}

\item{ani.text}{a logical value. If TRUE text on bars will be animated as well. If FALSE no text on bars will apear.}

\item{width}{png width (only works if format="png")}

\item{height}{png height (only works if format="png")}

\item{point}{png point size (only works if format="png")}

\item{res}{png resolution px/in (only works if format="png")}

\item{bg}{a logical value. If TRUE background will be drawn.}

\item{bg.from}{background color starting color}

\item{bg.to}{background color end color}

\item{bg.increase}{a logical value. If TRUE background color will increase from bg.from to bg.to}

\item{axes}{a logical value. If TRUE axes will be drawn.}

\item{cex.axis}{text size of axis}
}
\description{
Creates an animation of a one dimensional frequency table or vector. Animation can be saved to a folder in several formats.
}
\examples{
x<-sample(paste("group",letters[1:5]),200,T)
ani.barplot(prop.table(table(x)))
}
