% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp3d.R
\name{bp3d}
\alias{bp3d}
\title{A barplot for 3 dimensional contingency tables}
\usage{
bp3d(
  x,
  z,
  y,
  main = "",
  ylab = "frequencies",
  xlab = "",
  col = heat.colors(length(levels(y)) + 2)[1:length(levels(y))],
  space = 0.1,
  factor_space = 3,
  left_space = 0,
  box = T,
  boxcol = c("white", "grey"),
  legend = TRUE,
  locator = "topright"
)
}
\arguments{
\item{x}{main categorical variable on x-axis}

\item{z}{categorical sub variable on x axis}

\item{y}{dependend categorical variable}

\item{main}{title}

\item{ylab}{y axis label}

\item{xlab}{x axis label}

\item{col}{colors of bar segments}

\item{space}{space between bars within groups}

\item{factor_space}{factor of space between groups}

\item{left_space}{left space of bars}

\item{box}{draw boxes behind goups of x}

\item{boxcol}{color of boxes behind groups}

\item{legend}{draw legend}

\item{locator}{location of legend: "topright", "right" "bottomright"}
}
\description{
draws the conditioned freuqencies of a categorial variable on two further categorial variables
}
\examples{
x <- sample(paste("time",1:4),240,T)
z <- rep(c("A","B"),each=120)
y <- paste("answer",sample(1:7,240,T))
bp3d(x,z,y,main="bp3d()",xlab="treatment group")
}
