% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIV.R
\name{CIV}
\alias{CIV}
\title{Confidence Intervall Violin Plot}
\usage{
CIV(
  x,
  y,
  alpha = 0.05,
  ylim = range(y) + c(-(range(y, na.rm = T)[2] - range(y, na.rm = T)[1])/5, (range(y,
    na.rm = T)[2] - range(y, na.rm = T)[1])/5),
  ylab = "",
  xlab = paste("|---| ", 100 * (1 - alpha), "\% - confidence intervall", sep = ""),
  main = "",
  pch = 1 + as.numeric(factor(x)),
  col = "blue",
  colpoints = "black",
  cex = 0.7,
  names = levels(factor(x)),
  jit = 0.2
)
}
\arguments{
\item{x}{a categorial variable (independend variable)}

\item{y}{a numeric variable (dependend variable)}

\item{alpha}{error probability of confidence intervall}

\item{ylim}{y axis limits}

\item{ylab}{y axis label}

\item{xlab}{x axis label}

\item{main}{title}

\item{pch}{point character}

\item{col}{mean confidence intervall color}

\item{colpoints}{point color}

\item{cex}{point size}

\item{names}{names of groups on x axis}

\item{jit}{jitter intensity}
}
\description{
draws observations in metric variable for multiple groups, the group means, the 1-alpha confidence intervalls of the means and a density estimator (violin) of the distribution for every group.
}
\examples{
x<-sample(paste("group",1:4),100,T)
y<-10+as.numeric(factor(x))*2+rnorm(length(x))
CIV(x,y,main="Confidence Intervall Violin Plot")
}
