% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph2.R
\name{igraph2}
\alias{igraph2}
\title{igraph2}
\usage{
igraph2(
  x,
  freq = FALSE,
  label.cex = 1.7,
  vertex.size = 5,
  split2words = FALSE,
  rm.punctuation = FALSE,
  lowerize = FALSE,
  stemming = FALSE,
  capitalize = FALSE,
  rm.stopwords = FALSE,
  rm.numbers = FALSE,
  tkplot = FALSE,
  seed = NULL,
  min.n = 1,
  min.freq = 0,
  max.freq = 1,
  layout = 1
)
}
\arguments{
\item{x}{a list with connected elements as vector}

\item{freq}{logical. If TRUE displays frequency of connections in edge labels}

\item{label.cex}{numerical. Node label size}

\item{vertex.size}{numerical. Size of verices}

\item{split2words}{logical. If TRUE each element is further split at spaces}

\item{rm.punctuation}{logical. If TRUE punctuation/spaces are removed}

\item{lowerize}{logical. If TRUE lowerize to unify elements}

\item{stemming}{logical. If TRUE enables stemming}

\item{capitalize}{logical. If TRUE capitalize to display}

\item{rm.stopwords}{logical. If TRUE uses stop word removal}

\item{rm.numbers}{logical. IF TRUE excludes numbers}

\item{tkplot}{logical. If TRUE enables interactive adjustment with tkplot}

\item{seed}{numerical. Set seed for reproducible graphs}

\item{min.n}{numerical. Absolute minimum of involvement per element}

\item{min.freq}{numerical. Relative minimum of involvement per element}

\item{max.freq}{numerical. Relative maximum of involvement per element}

\item{layout}{numerical. Select out of:
1: layout.fruchterman.reingold(g)
2: layout.random(g)
3: layout.kamada.kawai(g)
4: layout.circle(g)
5: layout.reingold.tilford(g)
6: layout1 <- layout.sphere(g)}
}
\description{
draw network graph from list input
}
\examples{
x<-list(
     c("Laura","Ingmar"),
     c("Peter","Renate","Ingmar","Andrea"),
     c("Nassim","Ingmar","Sergej"),
     c("Laura","Rike","Andra"),
     c("Marlene","Nassim","Christina","Sabine"),
     c("Bela","Ingmar","Mariola","Nassim"),
     c("Gloria","Kim","Olek","Bolek"))
igraph2(x,seed=2)
}
