% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Image2Video.R
\name{image2video}
\alias{image2video}
\title{Convert an image sequence to a video or animated gif}
\usage{
image2video(
  folder = getwd(),
  i.format = "png",
  container = "mp4",
  fps = 8,
  last = 2,
  loop = 0
)
}
\arguments{
\item{folder}{folder with image sequence}

\item{i.format}{image format (png, jpeg, pdf only working for gif conversion)}

\item{container}{movie container: mp4, mov, avi, gif}

\item{fps}{frames per second used}

\item{last}{length of last frame in seconds}

\item{loop}{loop gif (0=True,1=False)}

\item{cex}{text size}
}
\description{
requires ffmpeg for video conversion: https://www.ffmpeg.org/
}
\examples{
x<-sample(paste("group",LETTERS[1:5]),200,T)
ani.barplot(prop.table(table(x)),folder="ani.barplot",format="png")
image2video(folder="./ani.barplot",cont="mov",i.f="png")
}
