% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemplot.R
\name{itemplot}
\alias{itemplot}
\title{itemplot}
\usage{
itemplot(
  x,
  col = "lightblue",
  col2 = c("red", "blue"),
  labels = c(levels(x)[1], levels(x)[length(levels(x))]),
  main = "",
  dec = 1,
  legend = TRUE,
  decmean = 2,
  quant.type = 4,
  pt.cex = 1.2,
  values = round(c(mean(a, na.rm = T), median(a, na.rm = T), quantile(a, 0.25, type =
    quant.type, na.rm = T), quantile(a, 0.75, type = quant.type, na.rm = T), length(a)),
    2),
  legendtext = c(as.expression(bquote("Mean" == .(round(mean(as.numeric(x), na.rm = T),
    decmean)))), as.expression(bquote("Median" == .(median(as.numeric(x), na.rm = T)))),
    as.expression(bquote("1. Quartile" == .(quantile(as.numeric(x), 0.25, type =
    quant.type, na.rm = T)))), as.expression(bquote("3. Quartile" ==
    .(quantile(as.numeric(x), 0.75, type = quant.type, na.rm = T)))),
    as.expression(bquote("N "[included] == .(sum(table(x))))),
    as.expression(bquote("N "[missing] == .(sum(is.na(x)))))),
  cex = 1
)
}
\arguments{
\item{x}{a likert scaled factor variable}

\item{col}{color of bars}

\item{col2}{color for mean and median}

\item{labels}{a vector of length 2 with low and high level labels}

\item{main}{title}

\item{dec}{decimal digits for percentage on bars}

\item{legend}{draw legend}

\item{decmean}{decimal digits for mean diplayed in legend}

\item{quant.type}{quantile type}

\item{pt.cex}{median and mean point size}

\item{values}{legend values}

\item{legendtext}{legends text}

\item{cex}{text size}
}
\description{
Function to draw distribution of Likert scaled item with relative frequencies, mean, 1., 2. and 3. quartile.
}
\examples{
x<-factor(sample(letters[1:6],100,T))
itemplot(x,main="itemplot()",labels=c("I fully\n disagree","I fully\n agree"))
}
