% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.line.R
\name{multi.line}
\alias{multi.line}
\title{multi.line}
\usage{
multi.line(
  x,
  y,
  group,
  main = "",
  axes = TRUE,
  xlab = "",
  ylab = "",
  xlim = range(x, na.rm = T),
  ylim = range(y, na.rm = T) + c(0, (max(y, na.rm = T) - min(y, na.rm = T))/10),
  points = TRUE,
  col = rainbow(length(levels(factor(group)))),
  pch = 1:length(levels(factor(group))),
  cex = 1,
  lines = TRUE,
  lty = 1:length(levels(factor(group))),
  lwd = 2,
  type = "line",
  locator = "top",
  horiz = TRUE,
  names = paste(levels(group), " ")
)
}
\arguments{
\item{x}{independend metric variable}

\item{y}{dependend metric variable}

\item{group}{grouping variable}

\item{main}{title}

\item{axes}{draw axes}

\item{xlab}{x axis label}

\item{ylab}{y axis label}

\item{xlim}{x axis limits}

\item{ylim}{y axis limits}

\item{points}{draw points}

\item{col}{point color}

\item{pch}{point character}

\item{cex}{point size}

\item{lines}{draw lines}

\item{lty}{line type}

\item{lwd}{line width}

\item{type}{draw regression line/s or lowess line ("line", "lowess", "both")}

\item{locator}{legend location}

\item{horiz}{align horizontally}

\item{names}{group names}
}
\description{
Draw multiple regression or lowess lines
}
\examples{
data(ChickWeight)
attach(ChickWeight)
Diet<-paste("Diet",ChickWeight$Diet)
multi.line(x=jitter(Time,2),y=weight,group=Diet,points=T,main="multi.line()",xlab="time",ylab="weight")
multi.line(x=jitter(Time,2),y=weight,group=Diet,points=F,main="multi.line() without points with lowess line",xlab="time",ylab="weight",type="lowess")
}
