% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orcid2name.R
\name{orcid2name}
\alias{orcid2name}
\title{orcid2name}
\usage{
orcid2name(x)
}
\arguments{
\item{x}{vector that contains ORCID and/or names}
}
\description{
function to convert ORCID to name with orcid.org api (needs valid ORCID token for OAuth authentication)
}
