% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettybarplot.R
\name{prettybarplot}
\alias{prettybarplot}
\title{prettybarplot}
\usage{
prettybarplot(
  x,
  col = "standard",
  ylab = "",
  xlab = "",
  main = "",
  names = "",
  ylim = c(0, max(x) + max(x)/7),
  box = TRUE,
  space = 0.2,
  border = "black",
  text = TRUE,
  cex = 1,
  coltext = 1,
  pre = "n =",
  post = "\%",
  digit = 1,
  legend = TRUE,
  bg = TRUE,
  bg.increase = TRUE,
  bg.from = "grey",
  bg.to = "white",
  axes = TRUE,
  cex.axis = par("cex.axis"),
  cex.names = par("cex.axis")
)
}
\arguments{
\item{x}{a one or two dimensional frequency table}

\item{col}{color of bars}

\item{ylab}{y axis label}

\item{xlab}{y axis label}

\item{main}{title}

\item{names}{name of bars/bar groups (in 2 dimensional table)}

\item{ylim}{y-axis limits}

\item{box}{draw box around plot}

\item{space}{space between bars}

\item{border}{border color of bars}

\item{text}{draw text on bars}

\item{cex}{text size}

\item{coltext}{color of text on bars}

\item{pre}{pre text for absolute frequencies}

\item{post}{post text for relative frequencies}

\item{digit}{digits to draw percentual frequencies with}

\item{legend}{Logical. If TRUE draws legend}

\item{bg}{draw background}

\item{bg.increase}{top-down increase/decrease in background color}

\item{bg.from}{background starting color}

\item{bg.to}{background end color}

\item{axes}{draw axes}

\item{cex.axis}{size of bar labels}

\item{cex.names}{size of axis labels}
}
\description{
draws a barplot of a one or two dimensional frequency table with frequencies on top of bars and pretty background
}
\examples{
x<-sample(paste("group",1:5),2000,T)
y<-sample(c("car","bike","bus"),2000,T)
# one dimensional barplot
prettybarplot(prop.table(table(x)))
# two dimensional barplot
prettybarplot(prop.table(table(y,x),m=2))
}
