% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileline.R
\name{profileline}
\alias{profileline}
\title{Profile line (of multiple groups) in multiple items}
\usage{
profileline(
  matrix,
  group = 1,
  type = "Median",
  groupnames = levels(factor(group)),
  main = "",
  col = 2:(length(unique(group)) + 1),
  pch = 1:length(levels(factor(group))),
  lty = 1:length(levels(factor(group))),
  lwd = 2,
  grid = TRUE,
  space_left = max(nchar(names(matrix)))/3 + 2.5,
  xlab = "",
  labels = "",
  extreme = FALSE,
  cex.axis.x = 1,
  cex.axis.y = 1,
  axis.top = TRUE,
  axis.bottom = TRUE,
  low_label = "",
  high_label = "",
  padj = 3,
  bottom.adj = 0,
  legend = TRUE,
  N_legend = TRUE,
  cex.legend = 1,
  title = TRUE
)
}
\arguments{
\item{matrix}{a data.frame or matrix containing the likert variables to draw}

\item{group}{grouping variable for profile lines (optional)}

\item{type}{type of line: one out of "Median", "Mean", "1. Quartile", "3. Quartile"}

\item{groupnames}{names of groups}

\item{main}{title}

\item{col}{color of group lines}

\item{pch}{point character}

\item{lty}{point character}

\item{lwd}{line width of profile line}

\item{grid}{draw grid}

\item{space_left}{space to y axis of first bar}

\item{xlab}{x axis label}

\item{labels}{labels displayed on x axes (must be of length}

\item{extreme}{only draw poles of labels (senseful for likert items)}

\item{cex.axis.x}{cex of x axis}

\item{cex.axis.y}{cex of y axis}

\item{axis.top}{draw top axis}

\item{axis.bottom}{draw bottom axis}

\item{low_label}{lowest level}

\item{high_label}{highes level}

\item{padj}{vertical adjust of low/high levels}

\item{bottom.adj}{mar bottom adjust}

\item{legend}{draw legend}

\item{N_legend}{draw legend of N observations per group}

\item{cex.legend}{character expansation of legend}

\item{title}{legends title}
}
\description{
draws a profile line of means or medians in multiple items and groups
}
\examples{
matrix<-data.frame(matrix(sample(-3:3,1000,T),ncol=20))
names(matrix)<-paste("Item",1:20); g<-paste("Group",rep(1:3,length=200))
for(i in 1:dim(matrix)[2]){matrix[,i]<-factor(matrix[,i],-3:3,c("fully disagree","2","3","4","5","6","fully agree"))}
profileline(matrix,g,main="",N_legend=T,pch="",type="Mean",ex=T,low="low",high="high",xlab="agreement")
}
