% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldmap2.R
\name{worldmap2}
\alias{worldmap2}
\title{worldmap2}
\usage{
worldmap2(
  country,
  connections = TRUE,
  region = NULL,
  main = "country and country connection frequency",
  ocean = "grey95",
  missingCol = "white",
  legend = TRUE,
  xpd = TRUE,
  mar = c(0, 0, 0, 0),
  mai = c(0, 0, ifelse(legend == TRUE, 0.1, 0), 0),
  palette = "Blues",
  legend1adjust = 0,
  legend2adjust = 0
)
}
\arguments{
\item{country}{a list with vectors of country names}

\item{region}{one of: ('eurasia' 'africa' 'latin america' 'north america' 'uk' 'oceania' 'asia')}

\item{main}{title}

\item{ocean}{ocean color}

\item{missingCol}{color of non present countries}

\item{legend}{logical. If TRUE legends are drawn}

\item{xpd}{logical. If TRUE ocean color is set to background color}

\item{mar}{mar adjustment}

\item{mai}{mai adjustment}

\item{palette}{colour palette from brewer.pal(). 
One of 'Blues', 'BuGn', 'BuPu', 'GnBu', 'Greens', 'Greys', 
'Oranges', 'OrRd', 'PuBu', 'PuBuGn', 'PuRd', 'Purples', 
'RdPu', 'Reds', 'YlGn', 'YlGnBu', 'YlOrBr', 'YlOrRd', 
'BrBG', 'PiYG', 'PRGn', 'PuOr', 'RdBu', 'RdGy', 'RdYlBu', 
'RdYlGn' 'Spectral'}

\item{legend1adjust}{adjust y position of upper legend and title}

\item{legend2adjust}{adjust y position of lower legend}

\item{connection}{logical. If TRUE draws country connections}
}
\description{
worldmap2
}
\examples{
worldmap2(country=list(c("Germany","Austria","S. Korea"),
c("Germany","Austria","S. Korea"),c("Germany","Austria","S. Korea"),
c("Germany","Spain"),c("Germany","Spain"),c("United States","Germany"),
c("Germany","Cuba","Bolivia","South Africa")),
          connection=T,legend=T,xpd=F,
          legend1=-10,legend2=-15)
}
