% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catplot.R
\name{catplot}
\alias{catplot}
\title{catplot}
\usage{
catplot(
  x,
  y,
  margin = 1,
  main = "",
  xlab = "",
  ylab = "",
  cex.lab = 1,
  marginleft = 5,
  cexmax = 4
)
}
\arguments{
\item{x}{multi categorial variable}

\item{y}{multi categorial variable}

\item{margin=1}{conditioned frquencies (1: cond. on y, 2: cond. on x, NULL: unconditioned)}

\item{main=""}{main title}

\item{xlab=""}{x axis label}

\item{ylab=""}{y axis label}

\item{marginleft=5}{left margin space}

\item{cexmax=2}{max size of squares}
}
\description{
Draws conditioned or unconditioned frequency table of two multi categorial variables
}
