% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ORCiD}
\alias{ORCiD}
\title{ORCIDs and the corresponding author names}
\format{
A data frame with 1,750,000 rows and 2 variables:
\describe{
  \item{ID}{ORCID}
  \item{Author}{Author name}
  ...
}
}
\source{
\url{https://orcid.org/}
}
\usage{
ORCiD
}
\description{
A dataset containing 1,750,000 ORCIDs and corresponding names
}
\keyword{datasets}
