% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_stacked.area.chart.R
\name{stacked.area.chart}
\alias{stacked.area.chart}
\title{stacked area chart}
\usage{
stacked.area.chart(
  x,
  col = (dim(x)[2] + 1):2,
  xpos.text = dim(x)[1] + 0.1,
  reorder = FALSE,
  xlab = "",
  ylab = "",
  cex = 0.9,
  xlim = c(1, length(rownames(x))),
  ylim = c(0, max(mat))
)
}
\arguments{
\item{x}{2 dimensional frequency table}

\item{col}{color of areas}

\item{xpos.text}{x position of text}

\item{reorder=FALSE}{Logical. If TRUE reorders by colsums of x}
}
\description{
stacked area chart
}
